capital_weather/requirements.txt

numpy==2.2.3
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
SQLAlchemy==2.0.39
urllib3==2.3.0
